genAI-labs/observability/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-logging==0.45b0
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-boto3sqs==0.45b0
structlog==24.1.0
python-json-logger==2.0.7
boto3==1.34.132
requests==2.32.3
pydantic==2.7.0
prometheus-client==0.20.0
aioboto3==12.4.0
